Allow enabling mTLS for memcached
This commit reworks the TLS section in order to add two new parameters:
1. 'SslVerifyMode' allows users to enable/enforce mTLS for memcached
2. 'AuthCertSecret' sets the secret name containing the certificate
   that clients should use.

See https://docs.memcached.org/features/tls/

By default SslVerfyMode is set to "None" or 0, but it can be set to "Request"
or "Require".

If set to "Require" it will:
1. prevent the creation of the non-tls memcached svc port (11211)
2. enforce client side certification validation

In addition, we add a new status bool (MTLS) to make it explicit
that mtls is used, and a new function GetMemcachedMTLSSecret() to
give services a way to fetch and use the required certificate.