Add TAGGED_AGGREGATES parameter and cleanups
This commit adds a new parameter called TAGGED_AGGREGATES (false by
default).
This allows operators to enable instanceha only on aggregates marked
with the EVACUABLE_TAG metadata, and also to disable/enable this
functionality without having to update aggregates.

We also took the chance to clean up a few pieces of code and added a
more explicit and informative warning message if a non-tagged compute
has failed.